refactor: divide pages on components
