+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { ThemeProvider } from '@mui/material/styles';
+import { ThemeProvider } from "@mui/material/styles";
 import {
   CssBaseline,
   AppBar,
@@ -12,17 +12,17 @@ import {
   ListItemText,
   Box,
   useMediaQuery,
-  IconButton
-} from '@mui/material';
+  IconButton,
+} from "@mui/material";
 import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
 import { motion } from "framer-motion";
-import MenuIcon from '@mui/icons-material/Menu';
+import MenuIcon from "@mui/icons-material/Menu";
 import Dashboard from "./pages/Review/Dashboard";
 import ChatHistory from "./pages/ChatHistory/ChatHistory";
 import PerformanceMetrics from "./pages/PerformanceMetrics/PerformanceMetrics";
 import CustomMetric from "./pages/CustomMetric/CustomMetric";
-import HallucinationMetric from "./pages/CustomMetric/HallucinationMetric";
-import AdditionalAnalytics from "./pages/CustomMetric/AdditionalAnalytics"; // Подключаем новый файл
+import HallucinationMetric from "./pages/HallucinationMetric/HallucinationMetric";
+import AdditionalAnalytics from "./pages/AdditionalAnalytics/AdditionalAnalytics";
 import theme from "./theme";
 
 const App: React.FC = () => {
@@ -38,16 +38,17 @@ const App: React.FC = () => {
     return () => window.removeEventListener("scroll", handleScroll);
   }, []);
 
-  const toggleDrawer = (open: boolean) => (event: React.KeyboardEvent | React.MouseEvent) => {
-    if (
-      event.type === "keydown" &&
-      ((event as React.KeyboardEvent).key === "Tab" ||
-        (event as React.KeyboardEvent).key === "Shift")
-    ) {
-      return;
-    }
-    setDrawerOpen(open);
-  };
+  const toggleDrawer =
+    (open: boolean) => (event: React.KeyboardEvent | React.MouseEvent) => {
+      if (
+        event.type === "keydown" &&
+        ((event as React.KeyboardEvent).key === "Tab" ||
+          (event as React.KeyboardEvent).key === "Shift")
+      ) {
+        return;
+      }
+      setDrawerOpen(open);
+    };
 
   return (
     <ThemeProvider theme={theme}>
@@ -57,8 +58,8 @@ const App: React.FC = () => {
           position="fixed"
           elevation={elevated ? 4 : 0}
           sx={{
-            backgroundColor: elevated ? 'rgba(0, 0, 0, 0.6)' : 'transparent',
-            backdropFilter: elevated ? 'blur(6px)' : 'none'
+            backgroundColor: elevated ? "rgba(0, 0, 0, 0.6)" : "transparent",
+            backdropFilter: elevated ? "blur(6px)" : "none",
           }}
         >
           <Toolbar>
@@ -77,41 +78,97 @@ const App: React.FC = () => {
             </Typography>
             {!isMobile && (
               <>
-                <Button color="inherit" component={Link} to="/">Обзор</Button>
-                <Button color="inherit" component={Link} to="/chat-history">Чаты</Button>
-                <Button color="inherit" component={Link} to="/performance">Производительность</Button>
-                <Button color="inherit" component={Link} to="/custom-metric">Доп. метрика</Button>
-                <Button color="inherit" component={Link} to="/hallucination-metric">Галлюцинации</Button>
-                <Button color="inherit" component={Link} to="/additional-analytics">Аналитика</Button> {/* Новый маршрут */}
+                <Button color="inherit" component={Link} to="/">
+                  Обзор
+                </Button>
+                <Button color="inherit" component={Link} to="/chat-history">
+                  Чаты
+                </Button>
+                <Button color="inherit" component={Link} to="/performance">
+                  Производительность
+                </Button>
+                <Button color="inherit" component={Link} to="/custom-metric">
+                  Доп. метрика
+                </Button>
+                <Button
+                  color="inherit"
+                  component={Link}
+                  to="/hallucination-metric"
+                >
+                  Галлюцинации
+                </Button>
+                <Button
+                  color="inherit"
+                  component={Link}
+                  to="/additional-analytics"
+                >
+                  Аналитика
+                </Button>
               </>
             )}
           </Toolbar>
         </AppBar>
         <Drawer anchor="left" open={drawerOpen} onClose={toggleDrawer(false)}>
-          <Box sx={{ mt: 8, p: 2, width: "100%" }} role="presentation" onClick={toggleDrawer(false)}>
+          <Box
+            sx={{ mt: 8, p: 2, width: "100%" }}
+            role="presentation"
+            onClick={toggleDrawer(false)}
+          >
             <List>
-              <ListItem button component={Link} to="/"><ListItemText primary="Обзор" /></ListItem>
-              <ListItem button component={Link} to="/chat-history"><ListItemText primary="Чаты" /></ListItem>
-              <ListItem button component={Link} to="/performance"><ListItemText primary="Производительность" /></ListItem>
-              <ListItem button component={Link} to="/custom-metric"><ListItemText primary="Доп. метрика" /></ListItem>
-              <ListItem button component={Link} to="/hallucination-metric"><ListItemText primary="Галлюцинации" /></ListItem>
-              <ListItem button component={Link} to="/additional-analytics"><ListItemText primary="Аналитика" /></ListItem> {/* Новый пункт */}
+              <ListItem button component={Link} to="/">
+                <ListItemText primary="Обзор" />
+              </ListItem>
+              <ListItem button component={Link} to="/chat-history">
+                <ListItemText primary="Чаты" />
+              </ListItem>
+              <ListItem button component={Link} to="/performance">
+                <ListItemText primary="Производительность" />
+              </ListItem>
+              <ListItem button component={Link} to="/custom-metric">
+                <ListItemText primary="Доп. метрика" />
+              </ListItem>
+              <ListItem button component={Link} to="/hallucination-metric">
+                <ListItemText primary="Галлюцинации" />
+              </ListItem>
+              <ListItem button component={Link} to="/additional-analytics">
+                <ListItemText primary="Аналитика" />
+              </ListItem>
             </List>
           </Box>
         </Drawer>
-        <Box sx={{ mt: 8, p: 0, width: "100vw", height: "100vh", display: "flex", flexDirection: "column" }}>
-  <motion.div initial={{ opacity: 0, y: 30 }} animate={{ opacity: 1, y: 0 }} transition={{ duration: 0.7 }} style={{ flex: 1, width: "100%" }}>
-    <Routes>
-      <Route path="/" element={<Dashboard />} />
-      <Route path="/chat-history" element={<ChatHistory />} />
-      <Route path="/performance" element={<PerformanceMetrics />} />
-      <Route path="/custom-metric" element={<CustomMetric />} />
-      <Route path="/hallucination-metric" element={<HallucinationMetric />} />
-      <Route path="/additional-analytics" element={<AdditionalAnalytics />} />
-    </Routes>
-  </motion.div>
-</Box>
-
+        <Box
+          sx={{
+            mt: 8,
+            pr: 7,
+            pl: 7,
+            width: "100vw",
+            height: "99vh",
+            display: "flex",
+            flexDirection: "column",
+          }}
+        >
+          <motion.div
+            initial={{ opacity: 0, y: 30 }}
+            animate={{ opacity: 1, y: 0 }}
+            transition={{ duration: 0.7 }}
+            style={{ flex: 1, width: "100%" }}
+          >
+            <Routes>
+              <Route path="/" element={<Dashboard />} />
+              <Route path="/chat-history" element={<ChatHistory />} />
+              <Route path="/performance" element={<PerformanceMetrics />} />
+              <Route path="/custom-metric" element={<CustomMetric />} />
+              <Route
+                path="/hallucination-metric"
+                element={<HallucinationMetric />}
+              />
+              <Route
+                path="/additional-analytics"
+                element={<AdditionalAnalytics />}
+              />
+            </Routes>
+          </motion.div>
+        </Box>
       </Router>
     </ThemeProvider>
   );
