+++ frontend/src/index.css
@@ -24,7 +24,9 @@ a:hover {
   color: #535bf2;
 }
 
+
 body {
+  overflow-x: hidden;
   margin: 0;
   display: flex;
   place-items: center;
