+++ frontend/src/pages/Review/components/ChartSection/BarChart.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { Box, Typography } from "@mui/material";
 
 interface BarChartProps {
   data: number[];
@@ -9,51 +10,46 @@ const BarChart: React.FC<BarChartProps> = ({ data, labels }) => {
   const maxValue = Math.max(...data, 1);
 
   return (
-    <div
-      style={{
-        display: "flex",
-        alignItems: "flex-end",
-        justifyContent: "center",
-        height: "80px",
-        gap: "6px",
-        padding: "4px 8px",
-      }}
+    <Box
+      display="flex"
+      alignItems="flex-end"
+      justifyContent="center"
+      height="100px"
+      gap={1}
+      padding="8px"
     >
       {data.map((value, index) => (
-        <div
+        <Box
           key={index}
-          style={{
-            flex: 1,
-            textAlign: "center",
-            position: "relative",
-          }}
+          display="flex"
+          flexDirection="column"
+          alignItems="center"
+          flex={1}
         >
           {/* Столбик */}
-          <div
-            style={{
+          <Box
+            sx={{
               height: `${(value / maxValue) * 100}%`,
+              width: "20px",
               background: "linear-gradient(180deg, #9C27B0 0%, #7B1FA2 100%)",
               borderRadius: "4px",
               transition: "height 0.3s ease, transform 0.2s ease",
               cursor: "pointer",
+              "&:hover": {
+                transform: "scale(1.05)",
+              },
             }}
-            onMouseOver={(e) => (e.currentTarget.style.transform = "scale(1.05)")}
-            onMouseOut={(e) => (e.currentTarget.style.transform = "scale(1)")}
           />
           {/* Подпись */}
-          <div
-            style={{
-              fontSize: "0.75rem",
-              marginTop: "6px",
-              color: "#ccc",
-              whiteSpace: "nowrap",
-            }}
+          <Typography
+            variant="caption"
+            sx={{ marginTop: "6px", color: "#ccc", whiteSpace: "nowrap" }}
           >
             {labels[index]}
-          </div>
-        </div>
+          </Typography>
+        </Box>
       ))}
-    </div>
+    </Box>
   );
 };
